[DOR-997] Available host memory undercount fix

Closes DOR-997

See merge request machine-learning/dorado!1612

+++ dorado/utils/include/utils/memory_utils.h
@@ -5,7 +5,7 @@ namespace dorado::utils {
 
 inline constexpr size_t BYTES_PER_GB{1024 * 1024 * 1024};
 
-size_t available_host_memory_GB();
+double available_host_memory_GB();
 size_t total_host_memory_GB();
 
 }  // namespace dorado::utils
